+++ cookbooks/aws-parallelcluster-slurm/recipes/install.rb
@@ -23,3 +23,4 @@ munge 'Install munge' do
   action :setup
 end
 include_recipe 'aws-parallelcluster-slurm::install_slurm'
+include_recipe 'aws-parallelcluster-slurm::install_pyxis'